♻️ cope with new types and move starknet code to a dedicated module
