+++ marketplace-indexer/src/infrastructure/apibara/starknet/events/contribution/gate_changed.rs
