+++ marketplace-indexer/src/infrastructure/apibara/starknet/events/contribution/unassigned.rs
@@ -2,17 +2,17 @@ use super::{EventTranslator, FromEventError, StarknetTopics, Topics};
 use marketplace_domain::{ContributionEvent, Event as DomainEvent, HexPrefixedString};
 use starknet::core::{types::FieldElement, utils::get_selector_from_name};
 
-pub struct Validated;
+pub struct Unassigned;
 
-impl EventTranslator for Validated {
+impl EventTranslator for Unassigned {
 	fn selector() -> FieldElement {
-		get_selector_from_name("ContributionValidated").unwrap()
+		get_selector_from_name("ContributionUnassigned").unwrap()
 	}
 
 	fn to_domain_event(mut topics: Topics) -> Result<DomainEvent, FromEventError> {
 		let contribution_id: HexPrefixedString = topics.pop_front_as()?;
 
-		Ok(DomainEvent::Contribution(ContributionEvent::Validated {
+		Ok(DomainEvent::Contribution(ContributionEvent::Unassigned {
 			id: contribution_id.into(),
 		}))
 	}
@@ -21,34 +21,31 @@ impl EventTranslator for Validated {
 #[cfg(test)]
 mod test {
 	use super::*;
-	use crate::infrastructure::apibara::proto::TopicValue;
 	use rstest::*;
 
 	#[fixture]
 	fn apibara_event_data() -> Topics {
-		vec![TopicValue {
-			value: vec![
-				0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-				0, 0, 0, 12,
-			],
-		}]
+		vec![vec![
+			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+			0, 0, 12,
+		]]
 		.into()
 	}
 
 	#[rstest]
 	fn selector() {
 		assert_eq!(
-			get_selector_from_name("ContributionValidated").unwrap(),
-			<Validated as EventTranslator>::selector()
+			get_selector_from_name("ContributionUnassigned").unwrap(),
+			<Unassigned as EventTranslator>::selector()
 		);
 	}
 
 	#[rstest]
 	fn create_event_from_apibara(apibara_event_data: Topics) {
-		let result = <Validated as EventTranslator>::to_domain_event(apibara_event_data);
+		let result = <Unassigned as EventTranslator>::to_domain_event(apibara_event_data);
 		assert!(result.is_ok(), "{}", result.err().unwrap());
 		assert_eq!(
-			DomainEvent::Contribution(ContributionEvent::Validated { id: 12.into() },),
+			DomainEvent::Contribution(ContributionEvent::Unassigned { id: 12.into() },),
 			result.unwrap()
 		);
 	}
