+++ marketplace-indexer/src/infrastructure/apibara/starknet/events/project/lead_contributor_removed.rs
@@ -2,18 +2,18 @@ use super::{EventTranslator, FromEventError, StarknetTopics, Topics};
 use marketplace_domain::{Event as DomainEvent, HexPrefixedString, ProjectEvent, ProjectId};
 use starknet::core::{types::FieldElement, utils::get_selector_from_name};
 
-pub struct LeadContributorAdded;
+pub struct LeadContributorRemoved;
 
-impl EventTranslator for LeadContributorAdded {
+impl EventTranslator for LeadContributorRemoved {
 	fn selector() -> FieldElement {
-		get_selector_from_name("LeadContributorAdded").unwrap()
+		get_selector_from_name("LeadContributorRemoved").unwrap()
 	}
 
 	fn to_domain_event(mut topics: Topics) -> Result<DomainEvent, FromEventError> {
 		let project_id: u128 = topics.pop_front_as()?;
 		let contributor_account: HexPrefixedString = topics.pop_front_as()?;
 
-		Ok(DomainEvent::Project(ProjectEvent::LeadContributorAdded {
+		Ok(DomainEvent::Project(ProjectEvent::LeadContributorRemoved {
 			project_id: project_id as ProjectId,
 			contributor_account: contributor_account.into(),
 		}))
@@ -23,24 +23,19 @@ impl EventTranslator for LeadContributorAdded {
 #[cfg(test)]
 mod test {
 	use super::*;
-	use crate::infrastructure::apibara::proto::TopicValue;
 	use rstest::*;
 
 	#[fixture]
 	fn apibara_event_data() -> Topics {
 		vec![
-			TopicValue {
-				value: vec![
-					0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-					0, 0, 0, 0, 0, 12,
-				],
-			},
-			TopicValue {
-				value: vec![
-					0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-					0, 0, 0, 0, 0, 24,
-				],
-			},
+			vec![
+				0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 12,
+			],
+			vec![
+				0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 24,
+			],
 		]
 		.into()
 	}
@@ -48,17 +43,18 @@ mod test {
 	#[rstest]
 	fn selector() {
 		assert_eq!(
-			get_selector_from_name("LeadContributorAdded").unwrap(),
-			<LeadContributorAdded as EventTranslator>::selector()
+			get_selector_from_name("LeadContributorRemoved").unwrap(),
+			<LeadContributorRemoved as EventTranslator>::selector()
 		);
 	}
 
 	#[rstest]
 	fn create_event_from_apibara(apibara_event_data: Topics) {
-		let result = <LeadContributorAdded as EventTranslator>::to_domain_event(apibara_event_data);
+		let result =
+			<LeadContributorRemoved as EventTranslator>::to_domain_event(apibara_event_data);
 		assert!(result.is_ok(), "{}", result.err().unwrap());
 		assert_eq!(
-			DomainEvent::Project(ProjectEvent::LeadContributorAdded {
+			DomainEvent::Project(ProjectEvent::LeadContributorRemoved {
 				project_id: 12,
 				contributor_account: 24.into()
 			},),
