+++ marketplace-indexer/src/infrastructure/apibara/starknet/mod.rs
@@ -1,0 +1,4 @@
+mod events;
+
+mod bytes;
+use bytes::*;
